Clean up enum and enum value framework
