+++ Assets/Audiovisuals/Animation Controllers/Astronaut_PlayerState.asset
@@ -1,0 +1,20 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2c0d77878f94c48858dbe0723b2df9b7, type: 3}
+  m_Name: Astronaut_PlayerState
+  m_EditorClassIdentifier: 
+  enumValues:
+  - {fileID: 11400000, guid: 88ee0ca519b81418ca14f313581b81da, type: 2}
+  - {fileID: 11400000, guid: 67271bcd40fbf46aca0b9e2af9aca1a2, type: 2}
+  - {fileID: 11400000, guid: 04392809cd55d44eab8c9fec192c1774, type: 2}
+  - {fileID: 11400000, guid: 054c8ee9aac9642cc9e7924b4fac7fda, type: 2}
+  - {fileID: 11400000, guid: 8d53288a7f2644f3689d0df8beba0e1e, type: 2}
