+++ Assets/Audiovisuals/Animation Controllers/Astronaut_PlayerState.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c5c97e45244ef4665be8f4ac9cc292dc
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
