+++ Assets/Scripts/_Utils/StateMachine/StateEnum.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2c0d77878f94c48858dbe0723b2df9b7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
